+++ docarray/base_doc/__init__.py
@@ -12,12 +12,12 @@ __all__ = ['AnyDoc', 'BaseDoc', 'BaseNode']
 def __getattr__(name: str):
     if name == 'DocResponse':
         import_library('fastapi', raise_error=True)
-        from docarray.base_doc.doc_response import DocResponse
+        from docarray.base_doc.docarray_response import DocArrayResponse
 
         if name not in __all__:
             __all__.append(name)
 
-        return DocResponse
+        return DocArrayResponse
     else:
         raise ImportError(
             f'cannot import name \'{name}\' from \'{_get_path_from_docarray_root_level(__file__)}\''
